refactor policy builder to separate struct with private and public funcs and utils.
